fix(typings): fix typing for fields that have a default value
Currently, a field that is defined as ["null", "string"] with a default value null in avro will result in a typescript typing that is optional and has typing 'string | undefined | null'. This is not correct, the field should be required (not optional) and the resulting typing should be 'string | null'.
This commit fixes this