fix(comments): do not strip urls halfway
If there is a url (http://example.com) in the schema (for instance in a doc-field), the previous solution would strip away ://example.com and the preceeding quotation mark. The result would be broken JSON. So we add a negative lookahead to check if the // is preceeded by a colon.